refactor(template): move copy code to pkg
